refactor: rename TCAIssueReporter to ComposableIssueReporter
- Rename TCAIssueReporter to ComposableIssueReporter for better naming consistency
- Add deprecated typealias for backward compatibility
- Rename configureTCAReporting() to configureComposableReporting()
- Add deprecated method for backward compatibility
- Update file name to match the new class name
